fix: minor cleanup
