+++ packages/io-auth-n-identity-domain/src/domain/index.ts
@@ -1,2 +1,2 @@
-export * from "./entities/lollipop/index.js";
+export * from "./entities/index.js";
 export * from "./value-objects/index.js";
