+++ packages/io-auth-n-identity-domain/src/domain/value-objects/lollipop/__tests__/lollipop-jwk-base64url.value-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { LollipopPublicKeySchema } from "../lollipop-public-key.js";
+import { LollipopPublicKeySchema } from "../lollipop-public-key.value-object.js";
 
 const toBase64Url = (value: unknown): string =>
   Buffer.from(JSON.stringify(value)).toString("base64url");
