+++ packages/io-auth-n-identity-domain/src/domain/value-objects/lollipop/assertion-type.value-object.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
-export const AssertionType = z
+export const AssertionTypeSchema = z
   .enum(["SAML", "OIDC"])
   .describe("Assertion type, e.g. 'SAML' or 'OIDC'")
   .brand<"AssertionType">();
 
-export type AssertionType = z.infer<typeof AssertionType>;
+export type AssertionType = z.infer<typeof AssertionTypeSchema>;
