+++ packages/io-auth-n-identity-domain/src/domain/value-objects/lollipop/index.ts
@@ -1,5 +1,5 @@
 export * from "./assertion-type.value-object.js";
-export * from "./lollipop-public-key.js";
+export * from "./lollipop-public-key.value-object.js";
 export * from "./lollipop-assertion-ref.value-object.js";
 export * from "./lollipop-content-digest.value-object.js";
 export * from "./lollipop-method.value-object.js";
