+++ packages/io-auth-n-identity-domain/src/domain/value-objects/lollipop/lollipop-public-key.value-object.ts
